Refactor BaseThumbnailBySizeAndFormatLoader
